requirements.txt
paho-mqtt>=1.6.1
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1 